kalamAI/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
websockets==13.1
redis==5.2.1
pydantic-settings==2.7.0
pydantic[email]==2.10.4
httpx==0.28.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
